Move run() into a new class, Runner
