+++ package.json
@@ -6,6 +6,7 @@
   "dependencies": {
     "assume": "^1.4.1",
     "aws4": "^1.6.0",
+    "joi": "^10.2.2",
     "libxmljs": "^0.18.4"
   },
   "devDependencies": {
